requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
strawberry-graphql-django>=0.25.0
django-environ>=0.11.2
gunicorn==22.0.0
uvicorn[standard]==0.29.0
django-tinymce==5.0.0
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-django==4.12.0
faker==40.15.0
